refactor: extract OkHttp-based HTTP client with support for cancellable requests
